main: Rework sub states in STATE_TRIGGERING
The sub states in STATE_TRIGGERING have been updated to allow
for a dedicated wait state, and to make it easier to have
accurate intervals between samples.

Tests are updated, and we now also test that the correct interval
is maintained over time.

+++ app/src/cbor/cbor_helper.h
@@ -20,4 +20,4 @@
  */
 int get_update_interval_from_cbor_response(const uint8_t *cbor,
 					   size_t len,
-					   uint64_t *interval_sec);
+					   uint32_t *interval_sec);
